# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardStats {
  totalPatients: number;
  appointmentsToday: number;
  hospitalized: number;
  admissionsToday: number;
  emergenciesWaiting: number;
  consultationsToday: number;
  analysesToday: number;
  imagingToday: number;
  invoicesToday: number;
  revenueToday: number;
  bedOccupancyPercent: number;
}

export interface ConsultationChartPoint {
  name: string;
  consultations: number;
  rendezVous: number;
}

export interface AdmissionChartPoint {
  name: string;
  admissions: number;
  sorties: number;
}

export interface ServiceChartPoint {
  name: string;
  value: number;
}

export interface RecentPatient {
  id: number;
  name: string;
  age: number;
  fileNumber: string;
  service: string;
  registeredAt: string;
}

export interface PatientMedicalInfo {
  allergies: string[];
  chronicDiseases: string[];
  majorHistory: string[];
}

export interface PatientListItem {
  id: string;
  mpiId: string;
  fileNumber: string;
  internalNumber: string;
  firstName: string;
  lastName: string;
  gender: string;
  dateOfBirth: string;
  phone: string;
  bloodType?: string | null;
  status: string;
  syncStatus: string;
  isIncomplete: boolean;
  potentialDuplicate: boolean;
  createdAt: string;
  updatedAt: string;
  nationality: string;
  country: string;
  medical: PatientMedicalInfo;
  createdById: string;
  siteId: string;
  service: string;
}

export interface UpcomingAppointment {
  id: number;
  patientName: string;
  service: string;
  doctorName: string;
  scheduledAt: string;
  status: string;
}

export interface AppointmentPatient {
  id: string;
  firstName: string;
  lastName: string;
}

export interface AppointmentListItem {
  id: string;
  patientId: string;
  patient: AppointmentPatient;
  doctorId: string;
  doctorName: string;
  departmentId: string;
  departmentName: string;
  scheduledAt: string;
  duration: number;
  status: string;
  notes?: string | null;
}

export interface CreateAppointmentBody {
  patientName?: string;
  patientFirstName?: string;
  patientLastName?: string;
  doctorName: string;
  departmentName?: string;
  scheduledAt: string;
  duration?: number;
  notes?: string;
  status?: string;
}

export interface UpdateAppointmentStatusBody {
  status: string;
}

export interface AlertItem {
  id: string;
  type: string;
  title: string;
  detail: string;
  description: string;
  severity: string;
  category: string;
  isRead: boolean;
  createdAt: string;
  patientId?: string | null;
}

export interface MarkReadResult {
  id: string;
  isRead: boolean;
}

export interface MarkAllReadResult {
  success: boolean;
}

export interface ConsultationListItem {
  id: string;
  number: string;
  patientId: string;
  patientName: string;
  patientMpi: string;
  doctorId: string;
  doctorName: string;
  specialty: string;
  serviceId: string;
  serviceName: string;
  siteId: string;
  siteName: string;
  date: string;
  scheduledAt: string;
  startedAt?: string | null;
  endedAt?: string | null;
  duration?: number | null;
  type: string;
  origin: string;
  reason: string;
  status: string;
  syncStatus: string;
  createdAt: string;
  updatedAt: string;
  createdById: string;
}

export interface CreateConsultationBody {
  patientName: string;
  patientMpi?: string;
  patientId?: number;
  doctorName: string;
  specialty?: string;
  serviceName?: string;
  date?: string;
  type?: string;
  origin?: string;
  reason: string;
  status?: string;
  duration?: number;
}

export interface UpdateConsultationStatusBody {
  status: string;
}

export interface BedsSummary {
  occupied: number;
  free: number;
  cleaning: number;
  outOfService: number;
  total: number;
  occupancyPercent: number;
}

export interface OrStatus {
  totalRooms: number;
  available: number;
  occupied: number;
  prep: number;
}

export interface BloodBankSummary {
  totalBags: number;
  available: number;
  urgentRequests: number;
  expiringSoon: number;
}

export type MedicationItemStatus = typeof MedicationItemStatus[keyof typeof MedicationItemStatus];


export const MedicationItemStatus = {
  ok: 'ok',
  low: 'low',
  critical: 'critical',
  expired: 'expired',
} as const;

export interface MedicationItem {
  id: number;
  name: string;
  quantity: number;
  unit: string;
  lowStockThreshold: number;
  expiryDate: string | null;
  expiringSoon: boolean;
  status: MedicationItemStatus;
  createdAt: string;
}

export interface MedicationPage {
  data: MedicationItem[];
  total: number;
  page: number;
  pageSize: number;
}

export interface UpdateMedicationStockBody {
  /** @minimum 0 */
  quantity: number;
}

export interface MedicationStockUpdate {
  id: number;
  quantity: number;
}

export interface VehiclesStatus {
  total: number;
  inService: number;
  available: number;
  maintenance: number;
}

export type GetPatientsListParams = {
search?: string;
status?: string;
gender?: string;
bloodType?: string;
};

export type GetAppointmentsListParams = {
search?: string;
status?: string;
departmentId?: string;
};

export type GetConsultationsListParams = {
search?: string;
status?: string;
type?: string;
origin?: string;
doctor?: string;
specialty?: string;
dateFrom?: string;
dateTo?: string;
};

export type GetMedicationsParams = {
status?: GetMedicationsStatus;
search?: string;
page?: number;
pageSize?: number;
};

export type GetMedicationsStatus = typeof GetMedicationsStatus[keyof typeof GetMedicationsStatus];


export const GetMedicationsStatus = {
  all: 'all',
  ok: 'ok',
  low: 'low',
  critical: 'critical',
  expired: 'expired',
} as const;