# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: getAppointmentsList
      tags: [appointments]
      summary: Get full appointment list
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: departmentId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Full appointment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentListItem"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentListItem"
        "400":
          description: Invalid input
  /appointments/{id}:
    patch:
      operationId: updateAppointmentStatus
      tags: [appointments]
      summary: Update appointment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentStatusBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentListItem"
        "404":
          description: Not found
...
  /consultations:
    get:
      operationId: getConsultationsList
      tags: [consultations]
      summary: Get full consultation list
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: origin
          in: query
          schema:
            type: string
        - name: doctor
          in: query
          schema:
            type: string
        - name: specialty
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Full consultation list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsultationListItem"
    post:
      operationId: createConsultation
      tags: [consultations]
      summary: Create a new consultation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConsultationBody"
      responses:
        "201":
          description: Created consultation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationListItem"
        "400":
          description: Invalid input
  /consultations/{id}:
    patch:
      operationId: updateConsultationStatus
      tags: [consultations]
      summary: Update consultation status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConsultationStatusBody"
      responses:
        "200":
          description: Updated consultation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationListItem"
        "404":
          description: Not found
...
components:
  schemas:
...
    AppointmentPatient:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
      required:
        - id
        - firstName
        - lastName
    AppointmentListItem:
      type: object
      properties:
        id:
          type: string
        patientId:
          type: string
        patient:
          $ref: "#/components/schemas/AppointmentPatient"
        doctorId:
          type: string
        doctorName:
          type: string
        departmentId:
          type: string
        departmentName:
          type: string
        scheduledAt:
          type: string
        duration:
          type: number
        status:
          type: string
        notes:
          type: string
          nullable: true
      required:
        - id
        - patientId
        - patient
        - doctorId
        - doctorName
        - departmentId
        - departmentName
        - scheduledAt
        - duration
        - status
    CreateAppointmentBody:
      type: object
      properties:
        patientName:
          type: string
        patientFirstName:
          type: string
        patientLastName:
          type: string
        doctorName:
          type: string
        departmentName:
          type: string
        scheduledAt:
          type: string
        duration:
          type: number
        notes:
          type: string
        status:
          type: string
      required:
        - doctorName
        - scheduledAt
    UpdateAppointmentStatusBody:
      type: object
      properties:
        status:
          type: string
      required:
        - status
...
    ConsultationListItem:
      type: object
      properties:
        id:
          type: string
        number:
          type: string
        patientId:
          type: string
        patientName:
          type: string
        patientMpi:
          type: string
        doctorId:
          type: string
        doctorName:
          type: string
        specialty:
          type: string
        serviceId:
          type: string
        serviceName:
          type: string
        siteId:
          type: string
        siteName:
          type: string
        date:
          type: string
        scheduledAt:
          type: string
        startedAt:
          type: string
          nullable: true
        endedAt:
          type: string
          nullable: true
        duration:
          type: number
          nullable: true
        type:
          type: string
        origin:
          type: string
        reason:
          type: string
        status:
          type: string
        syncStatus:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        createdById:
          type: string
      required:
        - id
        - number
        - patientId
        - patientName
        - patientMpi
        - doctorId
        - doctorName
        - specialty
        - serviceId
        - serviceName
        - siteId
        - siteName
        - date
        - scheduledAt
        - type
        - origin
        - reason
        - status
        - syncStatus
        - createdAt
        - updatedAt
        - createdById
    CreateConsultationBody:
      type: object
      properties:
        patientName:
          type: string
        patientMpi:
          type: string
        patientId:
          type: number
        doctorName:
          type: string
        specialty:
          type: string
        serviceName:
          type: string
        date:
          type: string
        type:
          type: string
        origin:
          type: string
        reason:
          type: string
        status:
          type: string
        duration:
          type: number
      required:
        - patientName
        - doctorName
        - reason
    UpdateConsultationStatusBody:
      type: object
      properties:
        status:
          type: string
      required:
        - status
...