# lib/api-zod/src/generated/types/appointmentListItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentPatient } from './appointmentPatient';

export interface AppointmentListItem {
  id: string;
  patientId: string;
  patient: AppointmentPatient;
  doctorId: string;
  doctorName: string;
  departmentId: string;
  departmentName: string;
  scheduledAt: string;
  duration: number;
  status: string;
  notes?: string | null;
}