# lib/api-zod/src/generated/types/consultationListItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConsultationListItem {
  id: string;
  number: string;
  patientId: string;
  patientName: string;
  patientMpi: string;
  doctorId: string;
  doctorName: string;
  specialty: string;
  serviceId: string;
  serviceName: string;
  siteId: string;
  siteName: string;
  date: string;
  scheduledAt: string;
  startedAt?: string | null;
  endedAt?: string | null;
  duration?: number | null;
  type: string;
  origin: string;
  reason: string;
  status: string;
  syncStatus: string;
  createdAt: string;
  updatedAt: string;
  createdById: string;
}