# lib/api-zod/src/generated/types/createAppointmentBody.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAppointmentBody {
  patientName?: string;
  patientFirstName?: string;
  patientLastName?: string;
  doctorName: string;
  departmentName?: string;
  scheduledAt: string;
  duration?: number;
  notes?: string;
  status?: string;
}