# lib/api-zod/src/generated/types/createConsultationBody.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateConsultationBody {
  patientName: string;
  patientMpi?: string;
  patientId?: number;
  doctorName: string;
  specialty?: string;
  serviceName?: string;
  date?: string;
  type?: string;
  origin?: string;
  reason: string;
  status?: string;
  duration?: number;
}